Revert "LPS-202905 Generate the link ASAP to avoid performance issues"
This reverts commit f264a5b7c7b2feeab569fb884a60b5dc129d8299.
